Remove Encoder base class
