+++ modules/core/src/adapters/deck-adapter.js
@@ -43,7 +43,6 @@ export default class DeckAdapter {
     this.enabled = false;
     this.getProps = this.getProps.bind(this);
     this.render = this.render.bind(this);
-    this.preview = this.preview.bind(this);
     this.stop = this.stop.bind(this);
     this._deckOnLoad = this._deckOnLoad.bind(this);
     this._getViewState = this._getViewState.bind(this);
@@ -92,8 +91,8 @@ export default class DeckAdapter {
   /**
    * @param {typeof import('../encoders').FrameEncoder} Encoder
    * @param {import('types').FrameEncoderSettings} encoderSettings
-   * @param {(params: any) => void} onStop
-   * @param {(prevCamera: any) => void} updateCamera
+   * @param {() => void} onStop
+   * @param {(prevCamera: import('../keyframes').CameraKeyframes) => void} updateCamera
    */
   render(
     Encoder = PreviewEncoder,
@@ -110,20 +109,16 @@ export default class DeckAdapter {
       );
     }
 
-    const innerOnStop = params => {
+    const innerOnStop = () => {
       this.enabled = false;
       if (onStop) {
-        onStop(params);
+        onStop();
       }
     };
     this.shouldAnimate = true;
     this.videoCapture.render(Encoder, encoderSettings, this.scene.lengthMs, innerOnStop);
     this.scene.animationLoop.timeline.setTime(this.videoCapture.encoderSettings.startOffsetMs);
     this.enabled = true;
-  }
-
-  preview() {
-    this.scene.animationLoop.timeline.play();
   }
 
   /**
